fdo#58142: fix frame selection in SwPageFrm::GetCrsrOfst:
The function does not handle the case where a paragraph was hit but no
background frame was hit properly, which may or may not (depending on
the platform) cause it to return a position of a frame when that is
undesirable, such as when called from SwWrtShell::UnSelectFrm().

When UnselectFrm() leaves a frame selected pasting multiple pictures
leads to a crash.

(regression from b2eed0eea05a926d3ab9355b4d1f34541c9ec72d)

Change-Id: I3604825f654c523a37f47a9fc660647d8b3e7077