n#792778 DOCX import: parse group shapes in oox only
Previously textframes inside groupshapes were tried to be imported as
TextFrames, but then their addition to a GroupShape failed, so the text
simply ended up as a normal paragraph. Fix this by importing members of
groupshapes as drawinglayer objects, just like how the WW8 import does.

Also fix two testcases, which implicitely tested that the groupshape VML
element is ignored on import.

Change-Id: I1a9fba8a5fd532203a825e55b1d5996277ea12fa