SwUndoInserts should not use a SwPosition:
There is a problem here; to see it paste something 3 times, then undo
twice and close the document, there is a SwIndexReg assertion because
the SwPosition in the one SwUndoInserts points to a node in the Undo
nodes array that is removed by the dtor of the other SwUndoInserts.

This is because the Undo objects are destroyed from the outermost Redo
backwards, which is usually a good idea but does not work for
SwUndoInserts, which (as they currently are) must be destroyed in the
other order.

But with the previous change to only store whole paragraphs in
SwUndoSaveCntnt it is possible to replace the SwPosition here with
a SwNodeIndex, which points directly to the SwTxtNode and thus does
not care if the position of that node changes due to whatever order
other SwUndoInserts are removed.

Change-Id: I4f0cf308d26f6b2e5aaa8997951c03ae7b2f0951