Simplify Visual Studio and Windows SDK related configurability
We had too many obscure options for the MSVC build. The configury
logic tried to take into consideration Registry keys or file locations
that likely never happen with the compilers we suppport (2008, 2010 or
2012) or the Windows SDKs from the same era.

Now there is just an option --with-visual-studio that can be used to
specify which Visual Studio version to use in case several are
installed. It takes as parameter the "year" version, 2008, 2010 or
2012. (There is also --with-windows-sdk that takes the version number
like 7.1A or 8.0, but I expect that this option will not be needed.)

The code tries hard to use sane defaults in all cases.

It is quite likely that there are bugs in the new logic. Those will
have to be fixed once noticed. Hopefully the basic idea is sane,
though.

Change-Id: I0a53acd334d35cbf2cd2fbc76a38d636e0d0123d