do not bother trying to keep PCH and non-PCH dependencies correct together
When compiling using a PCH, includes included using the PCH usually don't
show up in compiler dependencies output, so the only the PCH has a dependency
on the include, and the source has only a transitive dependency. Switching
to non-PCH (e.g. ENABLE_PCH= on cmdline) means the dependency is lost.
Trying with the PCH timestamp was a nice try, but it can sometimes still
miss something. So whoever wants to switch needs cleaning first or
just know what they're doing.