fdo#53521 Crash when running BASIC's "shell()" with a non-existent program
The problem is that the process handle is freed regardless of whether osl_executeProcess was successful or not. Therefore, a segmentation fault would result, as a random memory address would be freed. This patch checks whether osl_executeProcess was successful. If it was, then it frees the handle; if it wasn't successful, the memory is not freed.

Change-Id: Ic3af07be06e7af7730a3bcb1680c659a88b66af3
Reviewed-on: https://gerrit.libreoffice.org/1530
Reviewed-by: Michael Stahl <mstahl@redhat.com>
Tested-by: Michael Stahl <mstahl@redhat.com>