UNO components have the "lib" prefix only on Android
That they do is a leftover hack from before we started using the
DISABLE_DYNLOADING thing on Android, it isn't actually necessary any
more. Earlier when the UNO components were actually dynamic libraries
on Android (too, as on desktop OSes), they *had* to have names
starting with "lib" or they would be silently skipped when
packaging.

Change-Id: I11a4689f316e1ffb640a5c3110c63296d2833a84