always propagate font changes
Otherwise e.g. informing about new fonts when loading a document
with embedded fonts doesn't work. I'm not quite sure why this
should be limited to only printers or pdf, so this may be temporary
until a problem shows up. There'a also a warning about refcount
in ImplFontCache::Invalidate(), which seems to be bogus, because
invalidating caches shouldn't have effect on anything except for speed,
but with the way font handling in vcl is designed I wouldn't be
surprised if ImplFontCache wasn't actually a cache.

http://lists.freedesktop.org/archives/libreoffice/2012-September/038163.html

Change-Id: Ib68c0c10a0f5340a532c5c59396ef2cfbc588b5f