fdo#59928: sw: fix mouse selection of fields:
SwEditWin::MouseButtonDown: the selection of fields was messed up by a
call to rSh.SetCursor near the bottom of the method. This used to be
fixed up by SwEditWin::MouseButtonUp previously, but that was broken
with commit ec9d841c7f2e66c524563c75584f3628d39b60ee.

To fix that move the field selection code to a place where it prevents
the other code from running, in the hope that it is not necessary.

Also it does not make any sense to select whole words when clicking on a
field, just selecting the field seems much better.

Change-Id: I8d604450789844b6a446cf3a35f62ed530d0328e