Fix busted multi-repeated column in evolution database view
In the evolution address book where we have real column names of e.g.
first_name, second_name and city. On parsing via
OSQLParseTreeIterator::appendColumns that creates some labels using those real
names but the evo XResultSet gives them proper labels of First Name, Second
Name and City the munge means that here we have e.g. just "City" as a label
because it matches, while the others don't

The symptoms are to load the evolocal.odb and of the 128 columns column 5 is
repeated until 128

This is all a horrible confusing mess. It seems safest to catch the
mismatch of column counts and throw away the partial list and force
a generate of a full list.

Change-Id: I1d6e2a282bdd43acac63c366eb2a9d029aa17878