Go through all ScEditCell instantiations and fix memory leaks.
Changed the signature of the constructor (one that clones the text object)
to take a reference instead of a pointer, to smoke out the callers that
use this constructor.  Went through all its callers and made changes to
either 1) pass ownership to the cell instance (if the text object uses
the SfxItemPool instance returned from ScDocument::GetEditPool()), or 2)
pass as const reference to make it clear that the instance will get cloned.

Change-Id: I669e066d4739536bf8d3b356186503dcdfa303b0