Use consistent terminology
Why "Choose a Computer" but "Add Server"? Let's always call them "computers"
in the UI.

And in fact I think the feature to add a computer manually is kinda pointless;
is it too much to ask people to have their machines discoverable? I doubt
keeping a computer not discoverable through Bluetooth thwarts a determined
attacker from finding it anyway. (I am not saying that just finding it would
mean compromising it.)

Change-Id: Ib8041a8a793d84bee24587adac582c5fad03d1be